+++ flask_admin/tests/sqla/test_postgres.py
@@ -232,3 +232,42 @@ def test_boolean_filters(
         assert "true_val_1" in data
         assert "false_val_1" not in data
         assert "false_val_2" not in data
+
+
+def test_multiple_delete_integer_pk(
+    app: Flask,
+    sqla_postgres_db_ext: T_ANY_SQLA_PROVIDER,
+    postgres_admin: Admin,
+    session_or_db: T_LITERAL_SESSION_OR_DB,
+) -> None:
+    with app.app_context():
+
+        class IntModel(sqla_postgres_db_ext.Base):  # type: ignore[name-defined, misc]
+            __tablename__ = "test_bulk_delete_int_model"
+            id = Column(Integer, primary_key=True, autoincrement=True)
+            name = Column(String(50))
+
+        sqla_postgres_db_ext.drop_all()
+        sqla_postgres_db_ext.create_all()
+
+        m1 = IntModel(name="a")
+        m2 = IntModel(name="b")
+        m3 = IntModel(name="c")
+        sqla_postgres_db_ext.db.session.add_all([m1, m2, m3])
+        sqla_postgres_db_ext.db.session.commit()
+
+        param = skip_or_return_session_or_db(sqla_postgres_db_ext, session_or_db)
+        view = CustomModelView(IntModel, param)
+        postgres_admin.add_view(view)
+
+        client = app.test_client()
+
+        rv = client.post(
+            "/admin/intmodel/action/",
+            data=dict(action="delete", rowid=[str(m1.id), str(m2.id)]),
+        )
+        assert rv.status_code == 302
+        assert sqla_postgres_db_ext.db.session.query(IntModel).count() == 1
+        model = sqla_postgres_db_ext.db.session.query(IntModel).first()
+        assert model is not None
+        assert model.id == m3.id